refactor: centralize TaskStatus enum
- Move TaskStatus enum from task.entity.ts to dedicated enums/ directory
- Create enums/task-status.enum.ts with proper documentation
- Update all imports across entity, DTOs, service, and types
- Add barrel export at enums/index.ts

Benefits:
- Decouples enum from database layer
- Enables easier sharing with frontend/shared packages
- Improves maintainability with single source of truth
- Prevents circular dependency issues
- Semantically correct (enums are domain concepts)

Documentation added:
- docs/architecture/enum-management.md
  - Explains centralization pattern
  - Provides migration guide
  - Lists best practices and anti-patterns
  - Includes verification checklist

Relates to task: Enum Management Review
