+++ apps/backend/src/taskeroo/enums/index.ts
@@ -1,0 +1,1 @@
+export { TaskStatus } from './task-status.enum';
