+++ apps/backend/src/taskeroo/enums/task-status.enum.ts
@@ -1,0 +1,12 @@
+/**
+ * Task status enum
+ *
+ * Represents the lifecycle states of a task.
+ * Centralized enum that can be shared across layers.
+ */
+export enum TaskStatus {
+  NOT_STARTED = 'NOT_STARTED',
+  IN_PROGRESS = 'IN_PROGRESS',
+  FOR_REVIEW = 'FOR_REVIEW',
+  DONE = 'DONE',
+}
